refactor: multiple transformers to one for each type and add fallbacks
