+++ lib/YamahaReceiver/Transformers/BaseTransformer.js
@@ -20,17 +20,28 @@ class BaseTransformer {
         return result['YAMAHA_AV'][this.zone][0];
     }
 
+    /**
+     * Method to get a variable from an object parsed by xml2js
+     *
+     * This method can also be called with:
+     * - dot notation for fetching nested attributes, e.g. "ElementA.ElementB"
+     * - multiple name values will return the first one that's found
+     *
+     * @param array
+     * @param name
+     * @returns {String|null|*}
+     */
     getAttributeFromXMLArray(array, name) {
-        if (typeof name === "array") {
+        if (Array.isArray(name)) {
             for (let i in name) {
                 let result = this.getAttributeFromXMLArray(array, name[i]);
 
                 if (result !== null) {
                     return result;
                 }
-
-                return null;
             }
+
+            return null;
         } else if (name.indexOf('.') !== -1) {
             let value = array,
                 parts = name.split('.');
@@ -45,7 +56,11 @@ class BaseTransformer {
                 }
             }
 
-            return Entities.decode(value);
+            if (typeof value === "string") {
+                return Entities.decode(value);
+            }
+
+            return value;
         } else {
             if (this.attributeExistsXMLArray(array, name)) {
                 if (typeof array[name][0] === "string") {
@@ -67,9 +82,9 @@ class BaseTransformer {
                 if (result !== null) {
                     return result;
                 }
-
-                return null;
             }
+
+            return null;
         } else if (name.indexOf('.') !== -1) {
             let value = array,
                 parts = name.split('.');
@@ -87,6 +102,7 @@ class BaseTransformer {
             if (typeof value === "string") {
                 return Entities.decode(value);
             }
+
             return value;
         } else {
             if (this.attributeExistsXMLArray(array, name)) {
