+++ lib/YamahaReceiver/Transformers/BasicStatusTransformer.js
@@ -1,9 +1,9 @@
 'use strict';
 
 const xml2js = require('xml2js');
-const BaseTransformer = require('../BaseTransformer');
+const BaseTransformer = require('./BaseTransformer');
 
-class StateTransformer extends BaseTransformer {
+class BasicStatusTransformer extends BaseTransformer {
 
     constructor() {
         super('Main_Zone');
@@ -109,7 +109,7 @@ class StateTransformer extends BaseTransformer {
     getSoundVideoResult(result) {
         return this.getStrictAttributeFromXMLArray(
             this.getBasicStatusResult(result),
-            'Surround.Program_Sel.Current'
+            'Sound_Video'
         );
     }
 
@@ -122,4 +122,4 @@ class StateTransformer extends BaseTransformer {
 }
 
 
-module.exports = StateTransformer;
+module.exports = BasicStatusTransformer;
