+++ lib/YamahaReceiver/Transformers/PlayInfoTransformer.js
@@ -1,9 +1,26 @@
 'use strict';
 
-const xml2js = require('xml2js');
-const BaseTransformer = require('../BaseTransformer');
+const xml2js = require("xml2js");
+const BaseTransformer = require('./BaseTransformer');
 
-class BasePlayInfoTransformer extends BaseTransformer {
+class PlayInfoTransformer extends BaseTransformer {
+
+    transform(xmlResponse) {
+        return this.transformPlayInfo(xmlResponse).then(args => {
+            let playInfo = args[0],
+                availabilityResult = args[1],
+                playbackInfoResult = args[2],
+                metaInfoResult = args[3];
+
+            playInfo.available = availabilityResult === 'Ready';
+            playInfo.playing = playbackInfoResult === 'Play';
+            playInfo.artist = this.getAttributeFromXMLArray(metaInfoResult, 'Artist');
+            playInfo.album = this.getAttributeFromXMLArray(metaInfoResult, 'Album');
+            playInfo.track = this.getAttributeFromXMLArray(metaInfoResult, ['Song', 'Track']);
+
+            return playInfo;
+        });
+    }
 
     transformPlayInfo(xmlResponse) {
         let defaultPlayInfo = {
@@ -55,8 +72,7 @@ class BasePlayInfoTransformer extends BaseTransformer {
             'Meta_Info'
         );
     }
-
 }
 
 
-module.exports = BasePlayInfoTransformer;
+module.exports = PlayInfoTransformer;
