+++ lib/YamahaReceiver/YamahaReceiverClient.js
@@ -1,14 +1,11 @@
 'use strict';
 
 const axios = require('axios');
-const xml2js = require('xml2js');
-const SurroundProgramEnum = require('./Enums/SurroundProgramEnum');
-const BasicStatusStateTransformer = require('./Transformers/BasicStatus/StateTransformer');
-const PlayInfoAirPlayTransformer = require('./Transformers/PlayInfo/AirPlayTransformer');
-const PlayInfoSpotifyTransformer = require('./Transformers/PlayInfo/SpotifyTransformer');
-const PlayInfoUSBTransformer = require('./Transformers/PlayInfo/USBTransformer');
-const InputEnum = require('./Enums/InputEnum');
 const Entities = new (require('html-entities').XmlEntities)();
+const SurroundProgramEnum = require('./Enums/SurroundProgramEnum');
+const InputEnum = require('./Enums/InputEnum');
+const BasicStatusTransformer = require('./Transformers/BasicStatusTransformer');
+const PlayInfoTransformer = require('./Transformers/PlayInfoTransformer');
 
 class YamahaReceiverClient {
 
@@ -142,7 +139,7 @@ class YamahaReceiverClient {
         return this.getZone().then(zone => {
             if (this.hasPlayInfo(zone)) {
                 return this.get(zone, '<Play_Info>GetParam</Play_Info>').then(xmlResponse => {
-                    return this.getPlayInfoTransformerByZone(zone).transform(xmlResponse);
+                    return new PlayInfoTransformer(zone).transform(xmlResponse);
                 });
             } else {
                 return Promise.resolve(playInfo);
@@ -169,7 +166,7 @@ class YamahaReceiverClient {
     getState() {
         return this.request('GET', this.getMainZone(), '<Basic_Status>GetParam</Basic_Status>')
             .then(xmlResponse => {
-                return new BasicStatusStateTransformer().transform(xmlResponse);
+                return new BasicStatusTransformer().transform(xmlResponse);
             });
     }
 
@@ -335,30 +332,6 @@ class YamahaReceiverClient {
     attributeExistsXMLArray(array, name) {
         return typeof array[name] !== "undefined"
             && typeof array[name][0] !== "undefined";
-    }
-
-    getPlayInfoTransformerByZone(zone) {
-        if (zone === this.getAirPlayZone()) {
-            return new PlayInfoAirPlayTransformer();
-        } else if (zone === this.getIPodUSBZone()) {
-            return new PlayInfoIpodUSBTransformer();
-        } else if (zone === this.getSpotifyZone()) {
-            return new PlayInfoSpotifyTransformer();
-        } else if (zone === this.getNetRadioZone()) {
-            return new PlayInfoNetRadioTransformer();
-        } else if (zone === this.getServerZone()) {
-            return new PlayInfoServerTransformer();
-        } else if (zone === this.getJukeZone()) {
-            return new PlayInfoJukeTransformer();
-        } else if (zone === this.getTunerZone()) {
-            return new PlayInfoTunerTransformer();
-        } else if (zone === this.getUSBZone()) {
-            return new PlayInfoUSBTransformer();
-        } else if (zone === this.getMusicCastLinkZone()) {
-            return new PlayInfoMusicCastLinkTransformer();
-        }
-
-        throw new Error('Could not find PlayInfoTransformer for zone "' + zone + '"');
     }
 }
